backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
flask-jwt-extended==4.5.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0 